Document handling of values incompatible with keys
Document how a record will be treated during query execution if a
property path in filter or ordering specification will not result in a
value that is a valid IndexedDB key.

This behavior is based on native IndexedDB's behavior - indexes do not
include records that have the given index's key path resolve to a value
that is not a valid IndexedDB key.